Add support for detecting HDR and HBD images. Extend histograms.
This adds a isHDR() method to PaintImages as well as exposes it
and the existing is_high_bit_depth() to DiscardableImageMap.

The existing histograms are extended to include HDR and HBD; their
expiration dates have been extended as well.

This is the first part of support for rendering these images as
RGBA_F16 when an HDR display is connected.

Bug: 960620, 1076568
Test: Updated unittests.
Change-Id: I6e57351947748f49a598d7bb77cb74b3f31f2273
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2293226
Commit-Queue: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: ccameron <ccameron@chromium.org>
Auto-Submit: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788140}