Create a no-op scheduler job for 'Win10 Tests x64 1803'.
'Win10 Tests x64 1803' was removed from the branches because it is an
FYI builder that is not the mirror of any try builders and is red on the
branches so it is a distraction for sheriffs. Removing it from the
branches causes 'Win x64 Builder' to fail on the branches when
attempting to trigger it. The no-op scheduler job will allow the trigger
to succeed.

Bug: 1105483
Change-Id: I6e04ba407a62f7cdee4279d06538e8b10148df6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2297506
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Michael Moss <mmoss@chromium.org>
Commit-Queue: Michael Moss <mmoss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788205}