Implement Display Layout Preview
This change implements preview for display layout. Preview indicators
will be shown whenever a display in the display layout section of the
display settings page is dragged around to denote the future layout of
the displays.

Note: This is a continuation of CL: crrev.com/c/2146921

Bug: 1070122
Change-Id: I74b1fe62fffbdb15e0d97141ce392478a588d84b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2163438
Commit-Queue: Bailey Berro <baileyberro@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788244}