[cc/metrics] Fix incorrectly tracked expected aggregated frame counts
The regression is due to a previous CL:
https://chromium-review.googlesource.com/c/chromium/src/+/2234063

If an impl frame caused no damage, We decremented the
aggregated_throughput.frames_expected only when it is
not expecting main. But it could happen that at ReportFrameEnd,
it is expecting main, but after that, the main frame
reports no damage.

Bug: 1099325
Change-Id: If9709341abed2d818907dee0a20d5e52da37151f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2269739
Commit-Queue: Xida Chen <xidachen@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788246}