[Paint Preview] Fix overscoll refresh crash
Looks like we finally got a good tombstone (see bug). Looks like the
crash is because we weren't checking for the proto to be null in
UpdateSkpPaths(). I've now fixed this.

The reason the proto is null has to be that the (de)compression step is
either corrupting or failing resulting in no proto being read. The fix
is to just not compress during the test. I think this shouldn't be an
issue in production as this test (de)compresses too quickly in
succession.

Bug: 1105164
Change-Id: I28eb3b08afc24de82477e1245c3cf5121e46509c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2297679
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788263}