media/gpu/vaapi_unittest: move sandbox init to main()
crrev.com/c/2292813 (in 86.0.4202.0) removed one of the multiple calls
to VaapiWrapper::PreSandboxInitialization(), concretely the one in the
main() function, leaving only one in the test SetUp() method. This
broke the graphics.VAAPIUnittest.* in the CQ eve.

This CL fixes that by leaving a single call to said static method, in
the main() function.

Bug: 1105335, 1040291
Change-Id: I0b1bfa14f8768c6111697f591b0c8ff120c46e83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2298263
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788364}