[media] Provide async notification for GVAF codec support APIs.
The existing query APIs IsDecoderConfigSupported() and
GetVideoEncodeAcceleratorSupportedProfiles() are fallible because the
list of supported profiles may not have been retrieved from the GPU
process yet.

This CL adds APIs to check if the supported profiles have been fetched
and to register to be notified when they have been.

Bug: 1047994, 1105258
Change-Id: Ib4b9cce8599c00218cb72912186a893a91a65f00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2249103
Commit-Queue: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Harald Alvestrand <hta@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788385}