PDF Viewer: Move annotations bar to a separate element
Move the annotations bar to a separate element.

Also delegate some logic related to the ink controller to this new
element from the pdf_viewer, instead of relying on events and
getters to pass data back from the viewer down to the annotations bar
and back. Specifically:
- Pass the ink controller to the annotations bar via polymer bindings
- Listen for annotation-undo-state events and call redo/undo directly
  from the annotations bar, instead of doing this in the viewer.
- Add a new loaded event in the ink controller, so that the annotations
  bar can listen for this event to update the annotation tool when the
  ink controller loads, instead of needing to access the tool from
  the pdf viewer.

Bug: 1100431
Change-Id: Ifb4219d1df321d84666091f3707686a544f96e64
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285479
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788398}