Pass destination on clipboard read
Added the ClipboardDataEndpoint destination trying
to read the clipboard data as a as a parameter for all
ui::Clipboard Read*() functions. This destination is intended
to be used only on ChromeOS and Test platforms.
This CL adds this plumbing. Implementation of filtering will
be done in following CLs.

Bug: 1096456
Change-Id: If689ed3e38fc92c0ec50c043bf67312e59aeaa55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2266901
Commit-Queue: Aya Elsayed <ayaelattar@google.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Nikita Podguzov <nikitapodguzov@chromium.org>
Reviewed-by: Darwin Huang <huangdarwin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788407}