[TTS] Update metrics for M85 experiments.
Various updates to metrics recording for Contextual Search that are
needed for the M85 experiments in progress.

Some of the metrics are user actions that track the selection and
how it is being adjusted. Existing actions that track manual adjusting
of the selection was being done in the wrong part of the code making
it not useful for the experiment that we need them for.

Histograms for system resolve requests and responses are also added,
along with a histogram for translation.  Sometimes a histogram is used
as a simple counter, but there doesn't seem to be a natural way to do
that so a boolean histogram is used (with a signal that may not be
very important).

The new metrics entries are all in our UMA file, and they present a
convenient bottleneck for future logging of selection adjustment for
Smart Selection since it needs to know if the selection was adjusted
by the system or the user.

Metrics additions:
UserAction: "ContextualSearch.SelectionEstablished"
UserAction: "ContextualSearch.ManualRefineSingleWord"
UserAction: "ContextualSearch.ManualRefineMultiWord"
Histogram: "Search.ContextualSearch.SelectionExpanded"
Histogram: "Search.ContextualSearch.ResolveRequested"
Histogram: "Search.ContextualSearch.ResolveReceived"
Histogram: "Search.ContextualSearch.TranslationNeeded"

Also fix the logging of the CardTag.  A regression caused the mapping
from coca=>enum values to be called twice, which distorted the values.

BUG = 1102670, 966203

(cherry picked from commit f7fa8f0043240e13771568359204e5e87c5839c2)

Change-Id: I09f4941738e82c554f25ba9bfe0a7c417db9ba48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2286480
Commit-Queue: Brian White <bcwhite@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Auto-Submit: Donn Denman <donnd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787416}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2298109
Reviewed-by: Donn Denman <donnd@chromium.org>
Commit-Queue: Donn Denman <donnd@chromium.org>
Cr-Commit-Position: refs/branch-heads/4183@{#533}
Cr-Branched-From: 740e9e8a40505392ba5c8e022a8024b3d018ca65-refs/heads/master@{#782793}