Add tests for events metrics when commit is aborted
There are a few cases that the commit is aborted:
 1) LayerTreeHost is not visible;
 2) Main frame update is deferred;
 3) Commit is deferred;
 4) There is no update/damage.

In case (4), gathered events metrics are thrown away as handling of the
events had no effect. In cases (1) to (3), commit is deferred to a later
time and might happen in the future, so we should keep gathered events
metrics around to report in case the commit happened in the future. This
CL adds comments to explain this in code and adds tests to verify these
behaviors.

Bug: 1100491
Change-Id: I89c2bae0a78c0274a78cf9ff6c4979bbb53ec952
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285476
Commit-Queue: Mohsen Izadi <mohsen@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788480}