Add early return to NGOutOfFlowLayoutPart::Run()
Similar to what we have in NGOutOfFlowLayoutPart constructor
we should have an early return in NGOutOfFlowLayoutPart::Run()
when we have no positioned descendants.

We already have it a few lines later, but we can move it
to the beginning of the method to avoid any extra operation.

We also add a check for fragmentainer descendants to avoid
doing extra work if we don't have any.

BUG=1098231

Change-Id: I2d233c1c79a903d1696099c26029c2d09d59a728
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2279759
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Manuel Rego <rego@igalia.com>
Cr-Commit-Position: refs/heads/master@{#788556}