Cherry-pick of crrev/c/2289779
[Autofill Assistant] Allow to clear RequiredFields

This CL allows to clear a field instead of filling it. It will behave
the same way as filling a field in regards to the |forced| flag, i.e.
if the field is already empty (and we don't force it) we don't do
anything.

The field is, in the end, checked for emptiness. Non empty fields with
the |clear_value| flag will fail the action.

The |clear_value| flag should only really be used for <input> fields.
It can work with <select> as well (setting the value attribute) but
this is not guaranteed to actually empty the field (if the default has
a different value). The flag does not work in combination with
|fallback_click_element| at all.

(cherry picked from commit 164dbb35faf8ca6d15fc7a817627eba05a53172f)

Bug: 1105363
Bug: b/160850266
Change-Id: Ib170ee0f3e6931ad5f09a1664548c492e260e52a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2289779
Reviewed-by: Marian Fechete <marianfe@google.com>
Reviewed-by: Clemens Arbesser <arbesser@google.com>
Commit-Queue: Sandro Maggi <sandromaggi@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#787680}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2299221
Reviewed-by: Luca Hunkeler <hluca@google.com>
Cr-Commit-Position: refs/branch-heads/4183@{#549}
Cr-Branched-From: 740e9e8a40505392ba5c8e022a8024b3d018ca65-refs/heads/master@{#782793}