[devtools] Use idiomatic way to check for presence of member
Previously, we used the .get() method to get a pointer to an optional
mojo struct member. However, if the struct is inlined, this will return
a pointer even if the optional member is not present. This change
uses the more idiomatic way that relies on the operator bool in the
struct pointer to check for the presence. By doing it this way, new
code that copies the pattern automatically works (independent of
whether there are inlined mojo structs).

Bug: chromium:1064873
Change-Id: I08091e4347ee3cdf96b96e54c857ece72960bc0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2300082
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788607}