Move URLRequest::ReferrerPolicy enum into its own file.
This enum is part of the NetworkService API, and files outside of the
network service shouldn't need to transitively include a bunch of net/
classes to use it.

This CL also converts it from an enum to an enum class.

This CL is part of a chain of CLs that will reduce the 4000 build
steps in content_browsertests currently depending on url_request.h
to about 200.

Bug: 1103784
Change-Id: I3077442ba745fd4f4db64f110b7f8df39d751d00
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290929
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: David Van Cleve <davidvc@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788710}