[Paint Preview] Add player unit tests after refactoring
This CL adds separate unit tests suits for
- ScaleController
- ScrollController
- Viewport

BitmapState is tightly coupled to the mediator so that continues to be
covered by the mediator test. This CL also removes redundant mediator
tests and cleans up the BitmapState @VisibleForTesting annotation.

Many tests need to remain in the mediator to ensure that scroll/scale
interact correctly with bitmap updates.

Bug: 1099722
Change-Id: Ie38ece7d7e8bbbb05fa4356475f44ae537858139
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2295223
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Fred Mello <fredmello@chromium.org>
Reviewed-by: Mehran Mahmoudi <mahmoudi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788739}