Create PaymentUIsManager to take over PRImpl's UI logic
In order to reuse most of PaymentRequestImpl for WebLayer, to move the
business logic from PaymentRequestImpl to ComponentPaymentRequestImpl,
we need to put a clear cut between the UI logic and the business logic
within PaymentRequestImpl. So this CL is to create PaymentUIsManager to
take over the the UI logic of PRImpl.

Change:
* Move mUI into PaymentUIsManager and renaming as mPaymentRequestUI
* Move the PaymentUisShowStateReconciler into PaymentUIsManager

Bug: 1102522

Change-Id: I128a58c8a86c6325e408ad249571c45f127d44ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2299187
Commit-Queue: Liquan (Max) Gu <maxlg@chromium.org>
Reviewed-by: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788782}