Move download item separator rendering to background rendering.
Before the separator was rendered as a border, which had all sorts of
odd interactions with the focus ring (including being rendered at the
wrong time, causing it to not refresh immediately after the focus ring
disappears - see associated bug).

Now it is rendered as part of the background, which means it is behind
all other elements. This provides the same effect with less code and is
robust to focus ring changes.

Bug: 1049546
Change-Id: I472067bd28569fac8ef00a6d6b6335d5935b6e0a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2294342
Reviewed-by: Connie Wan <connily@chromium.org>
Commit-Queue: Dana Fried <dfried@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788787}