Replace some uses of TimeDelta::operator/() with equivalent forms.
Whether operator/() does integer or floating-point division is not
necessarily obvious to the reader (and will hopefully be changing in the
future).  Where it's at least as simple to get the same result without
using it, do so.

In many cases the code was effectively snapping to a multiple of some
time, which can be done in a shorter form using operator%().

Bug: 1104532
Change-Id: Id6614bfa91636bcf14568bb05212354b5a68c3e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2296859
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Timothy Dresser <tdresser@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788791}