[Autofill] Do not autofill user-edited fields.
This CL makes Autofill not fill fields whose fields have been
user-edited, except if their value is empty and the initial value
is empty as well. The rationale behind the exception is that if the
user removed a non-empty field value, Autofill should honour this.

There is an issue with TriggerRefill() causing fills with FormData
with empty values. These FormDatas are the result of cached
FormStructures converted back to FormData. This is addressed in a
follow-up CL.

Bug: 966873, 949094, 1075533
Change-Id: I7ab80dfa6edc85bf00b16ef5e99e777d032a9b7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2289410
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/master@{#788805}