[Offline indicator v2] Animate grid tab switcher using margins
crrev.com/c/2269197 added animations for the tab switcher using
translationY and avoided changing margins of the layout too often.

This was likely an over-optimization since the cards don't need to be
re-measured every time the margin is updated. This "optimization" also
caused a jump during the animation if the layout was scrolled all the
way to the down when the animation to hide the offline indicator
started.

This CL changes the animation to update the margins in every frame and
removes translationY from the properties.

Bug: 1104426
Change-Id: Ida177aa08d657b7d0f1535abf9b0ad709697aec1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2277075
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Cr-Commit-Position: refs/heads/master@{#788836}