[Autofill] Refill with respect to seen, not cached form.
This CL refactors AutofillHandler::OnFormsParsed() to take a vector
of FormData pointers instead of a vector of FormStructure pointers.
That way, refilling still has access to the seen FormData and does not
need to use the cached form for that purpose, which would mix up the
last-seen values with the initial values.

Bug: 966873, 949094, 1075533
Change-Id: Iba0e030b4dfa58edc05d125550e4a8ee56413f7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2291070
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Matthias Körber <koerber@google.com>
Cr-Commit-Position: refs/heads/master@{#788843}