[CBCM] Move ReportRequestQueueGenerator to components
Moves ReportRequestQueueGenerator from chrome/browser/enterprise to
components/enterprise so it can be reused in iOS code.

Note that the associated unit tests are harder to move because of their
extensive use of TestProfileManager, which is in chrome/browser. They
will be reworked and moved in a separate CL. crbug.com/1103732 has
been created to track that, with a TODO in the code.

Bug: 1091938
Change-Id: I24a2c2e258eb772e3910cdf99e8e1e7a83eb57a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290250
Reviewed-by: Owen Min <zmin@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#788848}