Link Doctor Removal Part 5: Actually Remove Link Doctor.
Link Doctor was a service that fetched suggestions on certain navigation
errors (e.g., typos in URLs, navigate to version in Google's cache,
etc). The feature has been disabled for some time now, and the
server-side service has been shut down. The disabled code is also using
an API that needs to be removed to deploy the backward/forward cache.
So it's time to remove the code.

This CL removes all the remaining Link Doctor code in Chrome.

TBR=estark@chromium.org (who actually signed off on it already)

Bug: 1030281
Change-Id: I7346ea116356d1955fd7774a6a0e945d3e1de249
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285324
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Edward Jung (EMEA) <edwardjung@chromium.org>
Reviewed-by: Nicolas Ouellet-Payeur <nicolaso@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788873}