Fix focus issue for TabGridDialog
http://crrev.com/c/2119188 removes the usage of PopupWindow in
TabGridDialog, which brings in a side effect that the dialog does not
grab the focus when it is opened. This leads to some focus shifting
issue in accessibility mode. This CL fixes this issue by manually
requesting the focus when dialog is opened and clearing the focus
when dialog is closed. Also, to fix the issue in accessibility mode,
we call sendAccessibilityEvent on top of requesting focus so that we
handle the focus transferring correctly.

Bug: 1101231
Change-Id: I42905fcb4e5c3f74de6e30f2f858bd66c4e14998
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2281088
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Wei-Yin Chen (陳威尹) <wychen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788924}