Propagate emulated window segments to renderer
Ensure that window_segments on WebDeviceEmulationParams are copied
over IPC and consumed in the renderer. The params are stored on
RenderWidgetScreenMetricsEmulator which uses these to call back to
set the emulated values on the RenderWidget. The original segments
are also stored so that they can be restored if the segments are no
longer emulated (either due to emulation being non longer applied
or if a future setDeviceMetricsOverride does not contain the segments).

Bug: 1099026

Change-Id: I9f32e432dc6e600ce3843e7774e7e0da53a97a5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2262193
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#788997}