[content] Export IsInactiveAndDisallowReactivation to content/public
Remove BackForwardCache::EvictIfCached which is being used
to check if the RenderFrameHost is in cache or not and evict it
accordingly.

For this, we already have IsInactiveAndDisallowReactivation,
which would be a more correct fit as we also check for other
RenderFrameHostImpl::lifecycle_state() for inactivity.

In this CL, we replace the usages in Autofill, PasswordManager
with IsCurrent() and in PermissionContextBase with
IsInactiveAndDisallowReactivation() which should only allow showing
UI when the frame is active.

BUG=1041021

Change-Id: I170fd671fb0c9c9c9979bbc2e78b0d8a9f18afa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2295819
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789016}