Trim in-memory logs in unsent_log_store when attempting to persist.
This will actually trim large logs if they cannot be uploaded right away. I'm not sure this is behavior we want, although I presume logs large enough to hit our limits (looks like 100K currently from metrics_reporting_service.cc) are large enough it probably makes sense to trim. Curious if you think this is risky.

After writing this, i considered changing the API such that the trimming call and persisting are seperate. Not sure which is better, but trimming at the time of persisting seems like a reasonable place to do it, so left it this way as it is simpler.


Bug: 1098317

Change-Id: Ieb72890d45f8ec3735403f5aa5695e6044df0b5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2257503
Commit-Queue: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789055}