[multiball] Fix DCHECK with FRE in multiwindow.
Fix a DCHECK that happens when two scenes are shown in FRE and the user
goes to background and then foreground.
The crash happens because the check for "is first run happening" was
only checking the current scene.
Instead, check all the scenes for FRE before scheduling the promo.
This way the attempt to show a promo isn't happening. The promo involves
a blocking UI, and the attempt happens in the second scene, hence the
crash.

Bug: 1104123
Change-Id: I2008894d4698edbb4ba9d2f19a45bae6c8bff8b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2292299
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789060}