[ios] Force target to be dirty if xcode_version changes
The binary XCTRunner will change if the version of Xcode changes,
but the timestamp and the input paths may not be reliable signals
to determine whether the target is dirty.

Force the command-line of build/config/ios/strip_arm64e.py to be
different if the version of Xcode changes. This will force ninja
to consider the target as dirty and will force using the correct
version of the binary (even if older than current output).

The parameter is ignored by the script, it is there only to force
ninja to consider the target dirty:

  $ ninja -C out/Debug-iphoneos -d explain
  ...
  ninja explain: command line changed for obj/build/config/ios/XCTRunner
  ...


Bug: 1101531
Change-Id: Iaf2baf8eadc2e5b3c66848d5f9be9638c5b44dc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2301975
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#789071}